requirements.txt
pillow==11.0.0
pytesseract==0.3.13
PyMuPDF==1.24.13
transformers==4.46.3
accelerate==1.1.1
tqdm==4.67.1